+++ src/IECoreHoudini/FromHoudiniGroupConverter.cpp
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (c) 2010-2012, Image Engine Design Inc. All rights reserved.
+//  Copyright (c) 2010-2013, Image Engine Design Inc. All rights reserved.
 //
 //  Redistribution and use in source and binary forms, with or without
 //  modification, are permitted provided that the following conditions are
@@ -95,7 +95,7 @@ FromHoudiniGeometryConverter::Convertability FromHoudiniGroupConverter::canConve
 	const GA_PrimitiveList &primitives = geo->getPrimitiveList();
 	
 	// are there multiple primitives?
-	size_t numPrims = geo->getNumPrimitives();
+	unsigned numPrims = geo->getNumPrimitives();
 	if ( numPrims < 2 )
 	{
 		return Admissible;
